BACKEND/apps/attendance/views.py
from rest_framework import viewsets, permissions, filters
from .models import Attendance
from .serializers import AttendanceSerializer
from apps.accounts.permissions import IsCompanyMember

class AttendanceViewSet(viewsets.ModelViewSet):
    serializer_class = AttendanceSerializer
    permission_classes = [permissions.IsAuthenticated, IsCompanyMember]
    filter_backends = [filters.SearchFilter]
    search_fields = ['employee__user__first_name', 'employee__user__last_name', 'date']

    def get_queryset(self):
        return Attendance.objects.filter(company=self.request.user.company)

    def perform_create(self, serializer):
        serializer.save(company=self.request.user.company)

    def perform_update(self, serializer):
        serializer.save(company=self.request.user.company)
