BACKEND/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
psycopg2-binary
pillow
cloudinary
django-filter
xhtml2pdf
python-decouple
celery
redis
whitenoise
gunicorn
openpyxl>=3.1.0
pandas>=2.0.0
